New feature - Jump-to-letter
The ability to jump to a specific letter in a media list. It will be hidden placed to the right of the current scrollbar, pressing to the right would bring up the scrollbar and pressing one more time would bring up the letter menu. Pressing select jumps straight to the first entry in the list beginning with the selected letter.